requirements.txt
Flask==2.2.3
gunicorn==20.1.0
google-cloud-texttospeech==2.14.1
google-cloud-storage==2.7.0